add Unitest to PointsAreNear
